Fixed logo issue on new installations
On new installations that BlueSpice Logo was missing because no default
was set on variable registration. This was done in BlueSpiceSkin which
makes no sense at all because the logo and and other images are not skin
related and would be missing when another is used.

Also made some improvements to the new installer. There is no more need
for renaming because no .template files are packaged.

 - Moved a couple of images and favicon to Foundation (not skin related)
 - Removed renaming of folders and files from installer
 - Added Author and changed some documentaion

Change-Id: Idf5b80c0a947ec9663d592d9b2fca9dd0c157d0b