BsFileSystemHelper: Quickfix to support NSFileRepo
This is just a quickfix to allow upload of namespace prefixed files. Such
file uploads do only appear in context with NSFileRepo extension.

For the future: It might be a good idea to have an additional parameter to
'uploadLocalFile' that allows to habe a target name that does not depend
on the real file name.

Change-Id: Ibe05998cf9d062400536676b5a3408fc322aca1b