Added missing resources
Since jenkins now validates on the existence of resources that were
referenced within CSS files we need to fix this.

Thos files are also missing in the original ExtJS Neptune Theme build. I
assume those rules are legacy code and not used anmore.

Still we want to make jenkins happy, so we add some dummy files.

Change-Id: I51b9c326357f101e419649e000a9a6d4d6af503f