ExtJS RL Module: Changed base class
This is necessary to have a working debug mode. If we derive from
ResourceLoaderFileModule we need to provide web accessible paths for the
client to load the script files directly. As we calcualte the language
file that get's embedded dynamically we'd need to duplicate this code to
provide a proper web path.

This means more effort than just to switch the base class and have no
loading of the static resource by the client.

PatchSet 2:
* Added target 'mobile'

Change-Id: I3fa78594532dfed020f2b935b5ac38e86ba21294