REL1_26 fixes/bs-filebackend-store: Removed meta
Removed the 'file_metadata' from 'BS.model.File' and the
'bs-filebackend-store' response. It contained data that made MW REL1_26
'ApiResult::applyTransformations' break.
As this field is not absolutely necessary and the information can be
fetched using a dedicated API the field got removed.
There is no known code depending on this field.

Change-Id: I3d6acbb7675d550674c4e29b6facca7c4f95067d