Fix for empty grid pages
ExtJS Grid/PagingToolbar: a user is on the last page of results
(bigger than "1") and deletes all entries from it.
When the store reloads with the old page info and sees an empty
result set it thinks there are no further results to display, even
if the 'totalProperty' says there are more. It just displays
"0 entries" to the user, leaving him without the change to select
a previous page.
Therefore we check for these circumstances and select a previous
page if possible

See ERM3303 for details

--- NEEDS CHERRYPICK TO MASTER ---

Change-Id: I0eb41ef31f8acdc58671c8623520f2804883736b