BsBaseTemplate et. al.: Serverside tab impl.
This change is due to a change in ResourceLoader between REL1_23 and
REL1_27. All scripts are now loaded async. There is no goot way to change
this.

Therefore we can not rely on jQueryUI being able to apply changes to the
DOM early enough. We need to set some stylings on the serverside to
overcome FOUC issues.

Change-Id: Ica46bc4ae54e2a7be15b21d7c7d594c3d81a8e35