ExtJS / RL: Removed obsolete "wait-for-css"
In the last version there was an issue with ExtJS code being executed
before ExtJS CSS was loaded completely. Since every calling code now uses
'Ext.onReady'-wrapper properly this is not an issue anymore.

In fact the implementation was more likely to cause issues as it
manipulates the state of 'ext.bluespice.extjs' directly.

This change removes the obsolete logic.

Change-Id: Ibd5412cb6106a22226167edc67900132ee157d58