Add function for retrieving custom params
Added a function that classes inheriting this base class can override and
pass any number and type of params to be passes when executing API
request.

Problematic is that there is already logic implemented that lets you pass
'query' param by overriding sQuery var. This is now unnecessary as 'query'
can also be passed over the new function, but for backwards compatibility
cannot be removed

Change-Id: I17ec9c0c6bbddc65daccdaca82a0bdcd736dc2f7