InsertFile: Fix for missing duplicate warnings
If you try to upload Test.pdf while already having Test.png on the wiki
there will be a warning, and this warning appears on Special:Upload, but
not on InsertFile upload dialog. There is only shows empty alert dialog.

In the case like normal check for duplicates does not retrieve anything
but the warning appears on upload action. But it does not appear in normal
"duplicates" node of response, but as "exists-normalized" and "exists"
attributes.

This kind of warnings also occures when getting normal duplicate warnings
but choosing not to ignore warnings.

Needs cherry-picking to REL1_27

Change-Id: I01175f1b926d82e9be6bcf0d3c5d6b744df036ca
ERM: #6495