RunJobsTrigger: Added dynamic interval system
Each IRunJobsTriggerHandler must not return it's own \Interval object.

Those objects can have a smart implementation when asked for the timestamp of
the next run.

E.g. Depending on the number of items to process, a handler can decide to run
more or less often.

Also it is easy to shift execution of handlers that register on the same
interval.

E.g. When there are three "OnceAday" handlers implementation ca decide to run
one at '1:00', the second one at '1:15' and the third one at '1:30'

Change-Id: I770db75aaa615269b09362f8539c984e4e48f889