Rename data folder to json
This extension contains a folder named "data" and a folder "Data" used
for namespacing.
On windows system this gives problems, because file names must be unique
in lowercase letters.
Renaming the "data" to json, because it just contains json files for
test data
Also move a file from folder "utility", because there is a namespace
"Utility", which have the same problem

Change-Id: I8ce19fa1aa156b270bcfef587fdf768da2cc787a