Fix for failed upload
Found that on some customers wfLoadFiles cannot retrieve the file
because not enough time has passed since the upload for it to be
discoverable.

In this case we dont need to really retrieve the file in order to
declare upload as success, successfull upload status is enough, and if
the file is available then do the rest.

Change-Id: Ifa186d829cd97d254191291beca8643e0c4cae4e
ERM: #9797