Improve searchresults when inserting link in VE
When during insertion of an internal link in visual editor
too many articles are found, the direct result could have
got lost before, for example when there where too many namespaces
or similar articles.

With this commit the entered Query is listed as first result
and the results are paginated.

Some superfluous sorting was removed. The sorting
functionality was redundant and did not work as expected.

ERM3585

Change-Id: I9f7e04736166cacd610f2a956812269585e9fa22