Add new generalized cache-subsystem
This adds a generic cache for titles as most of our extensions handle data
related to titles and do not use caching. I Added an example
implementation for pageprops. Old page prop mechanism is marked as
deprecated. I Also marked the DiscussionCout as deprecated, as it is not
used anywhere.

ERM:12692

Change-Id: I971017f9996fe1ba8761ce154f288b0f2ce9f61b