Add new generalized cache-subsystem
This adds a generic cache for titles as most of our extensions handle data
related to titles and do not use caching. I Added an example
implementation for pageprops. Old page prop mechanism is marked as
deprecated. I Also marked the DiscussionCout as deprecated, as it is not
used anywhere.

ERM:12692

Manual cerry-pick of https://gerrit.wikimedia.org/r/#/c/mediawiki/extensions/BlueSpiceFoundation/+/474740/

Change-Id: Ieb5dd4346d2ec4b99df7baab85714423b02c641f