Fix duplicate warning dialog opened twice
In Firefox the `change` event of the text field may occur twice. Therefore the
API call is made twice and evenually the warning dialog is created twice.

As the dialog has a specific ID it will crash when created the second time.

This is a manual cherry-pick of https://gerrit.wikimedia.org/r/#/c/mediawiki/extensions/BlueSpiceFoundation/+/492964/

NEEDS CHERRY-PICK TO `master`

ERM13780

Change-Id: Iad7ad871a893951f2c0ee6214143a4303affa304