Fix caching in PagePropHelper
A page may have no properties. This means an empty array is stored in the
cache. When getting the properties from the cache an empty array is completely
valid. We may only refresh the cache if the returned value was `false`.
Otherwise pages with no properties will cause a lot of write oprtations to the
cache on each request.

NEEDS CHERRY-PICK TO `master`!

Change-Id: I7084009bfa0716b867f74a80efceb0861418eab3