Create new Permissionrole: autocreateaccount
In order to be able to create an account on first login with
LDAP the group '*' needs to have the right 'autocreateaccount'.
With the new rolesystem this right was only granted to the role
'readers'.
With this change, the new role 'autocreateaccount' can be assigned
to the group '*' and autocreating for LDAP accounts can be enabled.

Cherrypick to REL1_31 recommended
ERM#14518

Change-Id: Icd4c742695de7a27dfb5784d9f988059561c0569