maintenance: Use Maintenance::countDown() instead of wfCountDown()
wfCountDown() was deprecated in MediaWiki 1.31 and Maintenance::countDown()
was also made available in this same version. This extension requires mw 1.34
so this replacement is safe.

NOTE: Leaving this line, though it's commented, maybe it will be used in the
nearest future? If so, then it will then be the correct code and no longer some
deprecated code.

Change-Id: Id202148c6a6ceaadcefcf7a4ab8360edd9fe3b0b