Change protection rights assignments to correspond with MW defaults
Well... kind of, "editprotected" by default should be given only to sysop,
so it makes sense to assign it to roles "admin" and "maintenanceadmin" only.
But "editsemiprotected" is more tricky, by default its given to "autoconfirmed",
but since we dont have corresponding role, i have assigned it to "editor".
However, this might not be what user expects when selecting
"Only allow autoconfirmed users" in protection form.
On the other hand, it might make sense, since "editor" is assigned by default
to "user" and therefore also to "autoconfirmed", so it would work as expected.

Added "editor" restriction level, since all restriction levels must be permissions,
added "editor" permissions, that is assigned to editor role.

So now user can select, protect for all, allow users in groups with "editor" role assigned,
or allow only administrators ("admin" and "maintenanceadmin" roles)

[3.1] - but this should be in 3.0.x as well, since permission assignments,
did not change from 3.0 to 3.1 (in this regard)

Change-Id: I15e3775aa1219a30243fe48779c942cc0a82ed29
ERM: #16167