Fix lockdown logic for searchable namespaces
If a namespace is locked down for any role, even if this role does not
contain the permission we are trying to lock, it will still deny the action
if the user does not belong to one of allowed groups.

This is wrong. We should only consider locking-down if any of the roles that
contain our permission is locked-down.

[3.1.x]

ERM 21391

Change-Id: Ifb9be7aaf563e56b959ac2ef2e1974795f8fd7d2