Fix broken call to User::isLoggedIn()
The method doesn't exist any more. It was just an alias for
isRegistered().

This patch also changes isAnon() to isRegistered(). This method is
not deprecated. However, isRegistered() is part of the much more
narrow UserIdentity interface. This change makes it much easier to
possibly replace usages of User with UserIdentity in the future.

Change-Id: I3cf663a9559329c5d4517c62dd8287d50d7e31f4
(cherry picked from commit ddbbcf6e6986f5b70f92b7a4768205fe546ed150)