Fix bs-usercombo setting value
In some cases, setting value to a UserCombo widget would not work.
This is true if value is set before store is finished loading, in which case
ExtJS itself will do some stuff in `checkValueOnChange` and reset the value.

I dont see any purpose of this function for UserPicker use-case, so i override
and kill its functionality

[4.2.x]

ERM 19054

Change-Id: Ie782e08867f2eff5a764d8c64a31fc8c4518ab37